New variable and behavior: fast_analyze_table for, well, faster ANALYZE TABLE
Introducing a new global/session variable: fast_analyze_table, a boolean that
defaults false. When the value is true (or 1), then an ANALYZE TABLE only
analyzes and persists stats for the clustering index (normally the PRIMARY KEY)
and skips the secondary keys.

The purpose of this change is to support faster ANALYZE TABLE operations, where
all we want is really to get an updated estimation of the number of rows in the
table. That information is extracted from the clustering index, not from the
secondary keys, and thus they can be skipped.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>