Only report unstable tests and don't fail the build (#84)
* Only report unstable tests and don't fail the build

* Show contents of log files after test failures

* Add a --flakes option

* Start a list of known flaky tests: flaky-tests.txt

* Add more flaky tests from #82

* One more flaky test

* Fix and check the path to flaky-tests.txt

* Ensure --flakes gets a non-empty arg

Oddly, `--flakes` is ignored, as is `--flakes=""`, because getopt doesn't
call the handler in either of those cases.

However: `--flakes ""` will now trigger the `die` added here.  The idea
here is to catch the error corrected in fe801e631d5a.

* Fix path for flaky tests file

This fixes the path for the flaky tests, fixes the logic for reading the
hash, decreases parallelism further and adds a bunch of tests to the
flakes set.

* Add flaky tests seen in #80

---------

Co-authored-by: Patrick Reynolds <patrick@piki.org>
Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>