Bug#35106114 ALTER table modify column default does not change native default
Problem:
Attempting to alter the default values value using inplace alter table
shows that the new value is not persisted in NDB dictionary.

Analysis:
Changing the default values value using inplace alter table is not
supported. There is already logic which attempts to check this
limitation. User wishing to change the default values value need to
perform a copying alter table.

Solution:
Change so that check for default values value change is always done for
columns which support and have a default value, thus detecting the
unsupported inplace change of default values value.

Change-Id: I57a95c4b5d9aa70b4e9928ae9a927c787bf09ae4