Bug#36259492 	InnoDB: `os_aio_func` is not working correctly for sync request
Symptom:
InnoDB Assertion failures happen if file opened with OS_FILE_NORMAL is used with `os_aio()`.

Root-cause:
In `os_aio_func()` when `WriteFile()` or `ReadFile()` return TRUE, meaning the IO request was executed synchronously, the `os_aio_func()` will check if number of bytes actually read/written is the same as
number of bytes to be read or written. If not, the IO slot in the array is released. But this happens without any latches, in particular we don't have array's mutex. When `WriteFile()` or `ReadFile()` return
TRUE, the event in the `OVERLAPPED` structure gets signaled either way, so in meantime the io_handler thread will pick this IO completion up, will postprocess the IO and release the slot.
This makes the `os_aio_func()` slot releasing to be racing with the io_handler thread and some thread will execute that for a second time, incorrectly.

Fix:
The `os_aio_func()` IO dispatching is rewritten to a loop (removed some gotos), where the pointer to slot is forgotten once we dispatch the IO correctly. The dispatch result code is simplified and in case `WriteFile()` or `ReadFile()` return TRUE we don't do anything but consider the IO as dispatched correctly.

Additional improvements:
Assertion in `AIO::release()` checking if the slot is reserved (that is, there is no double-release) is not debug-only now.
`AIO::reserve_slot()` is initializing values of `slot->n_bytes` and `slot->err` to empty (`0`, `DB_ERROR_UNSET`) values, so that they are not reused incorrectly in case of any potential bugs.

Change-Id: I05e32628543641136ad1a3789dfd5265a94d306f