Bug#36678321: Hypergraph: Sort-ahead doesn't provide row IDs to later sort that needs them [1/2, fix, noclose]
If a plan has a sort-ahead before a nested loop join to provide sorted
input to an operation higher up (for example DISTINCT), and there's
another sort higher up (for example ORDER BY), and the upper sort
requires row IDs, wrong results may be returned because the lower sort
doesn't provide row IDs.

The problem here is that the Filesort objects are created when
traversing the access path tree bottom-up in
FinalizePlanForQueryBlock(), but when creating a Filesort object for a
SORT access path somewhere down in the tree, it does not know if a
SORT access path higher up will require row IDs. This is not known
until the Filesort object for the path further up has been created,
and then it's too late to change the Filesort object in the lower path
to provide row IDs.

This patch fixes the bug by moving the creation of Filesort objects
out of the bottom-up processing of the tree and into a separate
top-down traversal step. Then it is known whether or not the Filesort
objects should be created with row IDs enabled.

Change-Id: Id08c1a3fc4ead03af0344fc7f6f830b812b2d900