WL#16416: Use NDV estimates from secondary engine in MySQL optimizer if available
Make the hypergraph optimizer use estimates on number of distinct
values (NDV) from secondary engine, if available. The NDV estimates
will be used when estimating the selectivity factor of an equality
between two columns, or to estimate the number of groups in a grouping
operation. The hypergraph optimizer can currently fetch this
information from either InnoDB index statistics or histograms, and the
following prioritized order will be used when deciding which
source of NDV estimates to use:

1. InnoDB index statistics
2. Statistics from secondary engine
3. Histograms

To make it possible for the hypergraph optimizer to fetch statistics
from a secondary engine, a new function pointer get_column_statistics
will be added to the handlerton. If a secondary engine implements this
function by adding content to a ha_column_statistics object, the
hypergraph optimizer may use this statistics during optimization.

A session variable, enable_secondary_engine_statistics, can be used to
control whether statistics from the secondary engine should be used.

Change-Id: I3fd990ce9fc0fe672f64a06b16a5c04b1b246934