Bug#36577747 	Scoped_heap incorrect usage may have memory leak
Symptom:
There is a possibility for memory leaks from a Scoped_heap incorrect usages.
One of such usages were fixed by:
Bug#35988311 	ASAN: memory leaks in bulk_load.merge-left-tree-taller

However, it would be nice if there is a mechanism that would make such errors less probable.

Root cause:
There are some use cases where a pointer to heap is needed to be stored as a local variable just to be passed as a reference (or a pointer to a pointer) to a method that may need the heap, and if the variable was nullptr, it would create one and change the variable by the reference.
In such cases, the heap gets disconnected with the Scoped_heap from which it seemed we took the pointer.

Fix:
One idea is to restrict calling `Scoped_heap::get()` to cases where we know it contains non-null pointer to heap.
This way we will not have a nullptr returned and replaced with a method that takes the variable with it as a pointer or reference.

Change-Id: I3c163a2c0b1213b4e2c43747494ec1badd9ae5b5