BUG#36729529 can_older_trx_be_still_active shouldn't base decision on trx_sys->rw_trx_list
The can_older_trx_be_still_active(max_trx_id) used two different
algorithms of checking if max_trx_id is larger or equal to any of active
transaction ids.

If the trx_sys->mutex was not congested it acquired it without waiting
and compared the max_trx_id to the minimum id in trx_sys->rw_trx_list.

This was wrong, because transactions are removed from this list *before*
they are officially releasing implicit locks - that happens later,
without trx_sys->mutex protection, when the trx is removed from
trx_sys->shards[].active_rw_trxs.

If the trx_sys->mutex was congested it fallen back to algorithm which
was correct, and didn't require this mutex: iterating over
trx_sys->shards[i].active_rw_trxs.peek().min_id() comparing it to
max_trx_id.

Side note: This was correct even though all store()s to the atomics
involved were memory_order_release, because it was assumed that the
max_trx_id comes from reading a page which had to happen after writing
to that page, which is sequenced after a transaction finishes
insert(id), thus the load() would see this value (or fresher), thanks
to ordering caused by page's latch x_unlock from thread that modified
the page happening-before all subsequent page reads that acquire this
latch (and in case of eviction: preads() happening-after pwrites()).

Unfortunately simply removing the first algorithm and only using the
second one was not very good for performance on idle server - iterating
over 256 shards takes more time, then acquiring uncongested mutex to
check one value. (10x slower in my testing).

Thus this patch, not only removes the first algorithm but also optimizes
the second, by maintaining atomic<trx_id> s_lower_bound value smaller
than all the id in the trx_sys->shards[].active_rw_trxs (if we only
consider those for which insert(id) happened before).
Note that this is very tricky. The simple part is to make it lower each
time we insert(id) which is lower. The difficult part is: how to ever
increase it, while somebody else can be inserting a smaller value in
parallel?

This patch proposes a solution in which a s_lower_bound_candidate value
is proposed from time to time and then atomically replaces the old
s_lower_bound. The process involves several steps, and is done by a
leader thread (selected exclusively by CASing s_lower_bound_candidate).
1. scan all shards to propose initial s_lower_bound_candidate value
2. scan again and update s_lower_bound_candidate again. Meanwhile other
   threads which are doing insert(id) should help by updating the
   s_lower_bound_candidate if id is smaller
3. finally replace s_lower_bound with s_lower_bound_candidate
Care must be taken in step 3, so that we accidentally don't "miss" an
id which was just inserted in parallel.
This is enforced by a protocol that insert(id) of a small id does 3
steps in the very specific sequence:
    m_min_id.store(trx_id);
    limit_to(s_lower_bound_candidate, trx_id);
    limit_to(s_lower_bound, trx_id);
While step 3. does a very specific CAS loop:
    while (true) {
      auto old_lower_bound = s_lower_bound.load();
      min_seen = s_lower_bound_candidate.load();
      if (s_lower_bound.compare_exchange_weak(old_lower_bound, min_seen)) {
        break;
      }
    }
Taken together this ensures that either the min_seen takes id into
account, or CAS will fail, or the s_lower_bound will be limited later by
insert(id).

This patch implements can_older_trx_be_still_active() which boils down
to O(1) comparison of max_trx_id to s_lower_bound most of the time.
In case max_trx_id is not smaller, we refine the s_lower_bound using
the 3 step procedure described above.
This solution is on par with trunk, yet hopefully correct as it no
longer uses the trx_sys->rw_trx_list.

Side node: the store to m_min_id is now using memory_order_seq_cst as it
is crucial to have store-load barier in insert(id). But this is the rare
case where shard was empty or even rarer case where id is inserted to
the shard out of order (this *can* happen because ids are assigned
independently from inserting them to the shards). The stores in
erase(id) are left memory_order_release as they can only make the value
larger, not smaller, which doesn't matter for correctness of lower bound.

Change-Id: Ibe4b6fc06e014837b475920b6f071f84e4eac709