Bug#36846564 Window function applied to BIT column produces non-BIT type (part 1/6 noclose)
As indicated by the reporter, the reason for the type change is the
fact that the result set goes through a materialization (tmp table),
which changes BIT to an integer type.

Solution: retain the bit type also for the tmp table. This has
repercussions, due to the weird encoding of bit type values inside the
null bytes of a record, so some methods for equality/comparison don't
work. We fall back on tmp table keys (of input record hash) when
record contains BIT field when doing DISTINCT. Note that there is
presedence for this approach, cf.
CreateTemporaryTableFromSelectList's comment:

  // temp_table_param.bit_fields_as_long is used to work around the
  // limitation of MEMORY tables not being able to index BIT
  // columns. But we also want to retain the type definition of
  // visible bit columns. So instead, force hash as the deduplication
  // method.

New tests are added to type_bit_<engine>.test to cover the different
scenarios where we needs to rely on tmp tables with BIT type in them.

Incidentally, the conversion of BIT to integer was introduced in
external bug#30324, resulting in the typing discrepancy mentioned in
the next paragraph.

This patch also makes the typing more consistent. Earlier, we had:

    create table t1(a bit(8));
    insert into t1 values (64),(65),(66),(66);
    mysql> select CAST(group_concat(a) AS CHAR) from t1;
    +-------------------------------+
    | CAST(group_concat(a) AS CHAR) |
    +-------------------------------+
    | @,A,B,B                       |
    +-------------------------------+
    1 row in set (0.01 sec)

    mysql> select CAST(group_concat(distinct a) AS CHAR) from t1;
    +----------------------------------------+
    | CAST(group_concat(distinct a) AS CHAR) |
    +----------------------------------------+
    | 64,65,66                               |
    +----------------------------------------+
    1 row in set (0.00 sec)

i.e., the encoding of the result changes when we add DISTINCT; instead
of concatenating the binary values, we would concatenate the
decimal value of those binary values which is slightly surprising.  We
now stick to the first form.

Change-Id: Iecd9be397e90571ab0004227422231edf864ed6c