Bug#36846564 Window function applied to BIT column produces non-BIT type (part 5/6 no-close)
remove GROUP_FLAG and Item::MARKER_BIT: no longer needed when we force hashing for BIT type tmp fields

This also fixes Bug#34683569 (external bug 108715): The problem was
the GROUP_FLAG survived to the client side, which interprets it as
a NUM flag, cf. the two identical values in the definition

 #define NUM_FLAG 32768             /**< Field is num (for clients) */
 #define PART_KEY_FLAG 16384        /**< Intern; Part of some key */
 #define GROUP_FLAG 32768           /**< Intern: Group field */

By removing the latter's usage, we no longer get that interference.
We also removed the no longer used variable group_null_item in
create_tmp_table.

The changes led to updating of canon for metadata.test, which used to
(wrongly) have the NUM flag set even though the datatype seen by the
client is MYSQL_TYPE_STRING, and also to several cases of shorter
result rows due to no longer using an extra byte for NULL with the
removal of

   share->reclength += group_null_items;

Change-Id: I7d9e865ec209d588e5c3450c8d7b65d0c3fb308b