Bug#32929873	REQUEST TO HAVE ADDITIONAL LOGGING MESSAGES IN GCS TO HELP IDENTIFY POSSIBLE UDF STUCK REASONS
Description
------------
While trying to workaround a problem while adding a new
member to the group by killing the thread in the host that
was running
"SELECT group_replication_set_communication_protocol('8.0.21')",
the thread has now been stuck with Command = Killed in the
processlist for over 40 minutes.

This bug purpose is to have GCS with increased/extra
information which might help to identify the reason why UDF
execution may cause GR to be stuck.

Analysis and Fix
------------
group_replication_set_communication_protocol is a group
action that demands that no traffic is flowing in the system
before upgrading or downgrading to a new protocol.

When this action is initiated, we:
- Block all sending of messages from every group member;
- We wait until all message flowing are delivered.

Most likely, the action gets stuck because the message
counter is not properly decremented, but we need extra
log mesage to prove this theory.

We have added new INFO messages on strategic places:
- When the action begins on all nodes;
- When we block sending message while waiting for the action
to end;
- When we receive messages and decrement the internal
counter.

Change-Id: Ib3a9405263965e7e5564c30215e0734686f222ba