Bug#37004689: Quantified comparison predicate and NOT NOT gives wrong result
Step 3: Fix the problem.

The problem manifests by giving wrong results for a WHERE clause like

  NOT NOT <op>ANY ( <subquery>)

where <op> is one of >, >=, <, <=, = or <>, and ANY may be replaced
with ALL. Exceptions are =ANY and <>ALL, which are usually transformed
into semi-join and anti-join.

The problem is with the implementation of
Item_func_nop_all::truth_transformer() and
Item_func_not_all::truth_transformer(): When a NOT modifier is applied,
these function will convert the operation into the inverse. Here we set
a new m_func by calling m_func_creator. The first application of this
replaces m_func with the identity. The second application inverts
the function. But this is wrong, it is only the object itself
(Item_func_nop_all or Item_func_not_all) and the m_all field that should
be inverted.

The fix is thus to remove the calls to m_func_creator().

Change-Id: I8f56f2fe220f5c4b2275e76acac85242b2ee7659