Bug#37014535: Assertion path->type == AccessPath::GROUP_INDEX_SKIP_SCAN || (path->type == AccessPath::FILTER && path->filter().child [2/2, fix]
The hypergraph optimizer sometimes raised an assertion when query
could use an index group skip scan.

This was caused by code which ensured that the estimated number of
rows returned was the same for the plans that used a group skip scan
and those that didn't. The code assumed that the plan under
consideration was either a group skip scan or a filter with a group
skip scan directly below. This assumption is not always correct. For
example, there could be more than one filter on top of the group skip
scan, or there could be a windowing step on top of the group skip
scan.

Fixed by propagating the row estimates down the tree in a more general
way, using a walker that doesn't care about the type of the paths in
the tree.

Change-Id: I716d130365d8442ef6458b7e3c4f98e0e0280dff