Bug#37022773 (1/2) Separate handling of NdbApi kernel and Api sourced errors
Add testcases to testScan showing problems with Api and Kernel
error interactions in scans.

  testScan -n ScanErrorCleanup T1

    Shows case where API error tramples kernel side error resulting
    in 'nocloseneeded' behaviour for scan, reuse and then TC state
    crash

  testScan -n ScanErrorHandling T1

    More comprehensive testcase covering combinatorial space of
    scan error sources :
      - Type : Table, Range, Range-ordered
      - Locking : Y/N
      - Delay before iteration (racing incoming signals or not)
      - When error occurs : Pre-exec, Pre-iterate, During-iterate, After
      - What error occurs : None, Early scan reject, Late scan fail, Api error

    Test checks :
      - (No crashes / no overall test timeout)
      - Number of API side Transaction objects does not increase
      - Subsequent transactions do not crash (kernel side object state ok)

Change-Id: I47a0733940c94f8e82c600b0242755b1cea68d31

Bug#37022773 (2/2) Separate handling of NdbApi kernel and Api sourced errors

Scan objects allow API interaction concurrently with the reception of
incoming signals related to the scan.

Incoming signals use the scan object's error code to manage scan
protocol error handling and subsequent closure.

Problem

This is flawed as the API may set a scan object's error code so that
subsequent closure does not operate correctly.

This can lead to :
 - Scan protocol timeout waiting for signals that the kernel will
   not send
 - Failure to close the scan resources on the kernel, leading to
   data node side state assertions.

Solution

The signal handling part of the code assumes that any error which
is set has been set by e.g. SCAN_TABREF handling.

Make this true by maintaining separate 'kernel error' and
'API/overall error' objects.

Signal handling sets only the kernel error object.
Signal handling based on the error examines only the kernel error
object
When the user makes a blocking call into the API (nextResult/close),
if the kernel error object is set, then it is copied to the API error
object so that the user becomes aware.

Refactoring

When implementing the fix, it can be seen that an analogous design is
used for SPJ, with the implementation encapsulated rather than implemented
in NdbTransactionScan.

This model is used to extend the 'execCLOSE_SCAN_REP()' method in a symmetric
way, which encapsulates the scan object's handling of errors and completions.

Change-Id: I8b2bc6471e6b4b826a543b51056dac755e25fcb6