Bug#37022901 (1/7) Improve NdbApi Table and OI scan protocol timeout handling
Add transactions_full base table to ndbinfo

The existing cluster_transactions table is filtered at source
in TC to exclude ApiConnectRecords which are seized, but not
in use.
This is likely aligned with the user's expectations, but makes
some resource usage unobservable.

A new base table is added without this filtering, as a
parameterised variant of the existing table.

No user-visible view is added at this time.

Change-Id: I23cff8a6ac0f719e82c843d465504929d7bc733a

Bug#37022901 (2/7) Improve NdbApi Table and OI scan protocol timeout handling

Add reduced timeout DBUG capability to NdbAPI.

New DBUG variable allows tests to be run to show timeout behaviour
without needing to wait 6 minutes for every iteration.

Modify direct access to the value from check_send_timeout() to ensure
that it is affected when the API timeout is reduced.

Change-Id: I7af78a6c70845323ee5e13af0e5b3a8f80791b02

Bug#37022901 (3/7) Improve NdbApi Table and OI scan protocol timeout handling

Add error inserts for testing timeouts

  5112 (LQH) : Range scans get an infinite sequence of MRR ranges so that
               they never terminate.
               With a non-matching pushed filter, they will scan indefinitely
               with no results

  8124 (TC)  : SCAN_FRAGCONF signals stall indefinitely, causing scans to
               timeout, and close() requests are unable to resolve until
               error insertion is cleared

Change-Id: Id7f7f29d2eab5a87e30cef40a02817d4355d5fda

Bug#37022901 (4/7) Improve NdbApi Table and OI scan protocol timeout handling

New MTR testcase ndb_scan_protocol_timeout

New testcase added which tests system behaviour of :
  - Unordered + Ordered scans
  - 'Load' and 'Bug' timeout scenarios (via error inserts)

Test checks :
  - TC resource usage to ensure no 'leaks' of ApiConnectRecords
  - Ability to query after all tests are complete

New include file ndb_get_api_connect_count uses new
ndbinfo.ndb$transactions_full base table to observe seized but
inactice (CS_CONNECTED) ApiConnectRecords.

No result file recorded as test fails without subsequent patches.

Change-Id: I42f920e98117f6f67a3122a2879047e8535a5738

Bug#37022901 (5/7) Improve NdbApi Table and OI scan protocol timeout handling

Remove error code 4008 handling from closeTransaction()

Scan operation variants may set error 4008 on both a 'scan' and 'main'
transaction object when a scan times out.
When that main transaction is closed in closeTransaction(), the presence
of error 4008 causes it not to be closed, but rather 'forgotten'.

This results in both the main transaction object + ApiConnectRecord, and
any related scan transaction objects + their ApiconnectRecords, to be
leaked.

This occurs whether or not the scan encountering 4008 has subsequently
been closed correctly at the kernel.

The leak is detected via  assertion failures in debug builds
(as NdbList detects that some allocated resource has not been returned).

As part of fixing timeout behaviour this special error code based
behaviour is removed.

If there is some situation where the kernel side of a scan cannot be
closed then the kernel side object may be leaked (until API disconnection),
but the api side object is not.

The baseline ndb_scan_protocol_timeout test result is recorded here,
showing 4 situations, each resulting in 1 ApiConnectRecord object being
leaked at TC.

Change-Id: I1ead0a95ebfdb2ab03276df437aa4a6cc0963bd7

Bug#37022901 (6/7) Improve NdbApi Table and OI scan protocol timeout handling

Protocol timeouts mark scans as needing close at kernel

The current timeout handling results in an error being set on the object,
but not in a way that ensures that an attempt will be made to cleanup
the kernel-side of the scan when close() is called.

The behaviour then depends on whether there were any 'confirmed' receivers
at the time the error was received (unlikely).

This is changed to handle an api protocol timeout in a similar way
to receiving a SCAN_TABREF from the kernel with an error code and
the 'need_close' flag set.

The code invoked in this case is reused.

This should result in a clean close of the scan in cases where the
scan is timing out due to 'LOAD' (e.g. the scan is busy but has returned
no results for 6 minutes).

In 'BUG' cases it will result in an attempt to close the scan (which may
help, depending on the problem).

The testcase result is updated to show that the 'LOAD' cases now have no
ApiConnectRecord leak for ordered scans.

Unordered scans continue to leak as they have a further bug

Future ideas :
  - Allow users to control timeout duration
  - Allow users to set a non-error timeout (polling)
  - Have kernel send a scan HB to reset api protocol timeout

Change-Id: I92aad65f6214505cd6abd5f3e97997ac20084c9f

Bug#37022901 (7/7) Improve NdbApi Table and OI scan protocol timeout handling

Only set ReleaseOnClose if close() times out

An NdbTransaction's ReleaseOnClose member is used to indicate that the
TC ApiConnectRecord associated with the NdbTransaction is 'gone', and so
when the transaction is next closed, the API side NdbTransaction object
should return to an anonymous free pool, ready to be associated with a
different ApiConnectRecord object.

The primary use case for this is when a data node fails - all the
ApiConnectRecords it had are now gone / unreachable, so the API must
detach its NdbTransaction objects from those.

Cases where current scan code is setting this member to true :

 - executeCursor()
   Node fail detected when sending : OK

 - In unordered scan nextResult() for 3 cases
   scan timeout (-1) : Maybe
   node failure (-2) : OK
   node seq change / signal send failure (-3) : Maybe

 - In close_impl()
   When node seq is changed : OK
   Wait-for-outstanding
     timeout (-1) : Maybe
     node failure (-2) : OK
   Send-close
     signal send failure : OK
     timeout (-1) : OK
     node-failure (-2) : OK

The ordered scan nextResult() code is not setting this member to true
as it assumes that in cases where the node fails, the sequence
number will change and this will be detected in close_impl()

This explains the test result difference between the ordered and
unordered scan cases.
If there is a timeout in nextResult() in an unordered scan the
nextResult() code sets theReleaseOnClose flag even in cases where
close_impl() manages to close the scan, resulting in an ApiConnectRecord
leak.

This patch aligns the unordered nextResult() code with the ordered
nextResult() code :
  - There is no explicit setting of theReleaseOnClose flag
    for timeout *or* node failure
  - The change in node sequence number is relied on to detect
    the node failure case
  - Extra logging is added so that if there is a timeout case,
    it is logged as either :
    - Timeout occurred and was successfully handled in close()
    - Timeout occurred, and close failed to handle.

The ndb_scan_protocol_timeout testcase result is updated to show
the removal of the unordered scan 'LOAD' scenario ApiConnectRecord
leaks with this patch.

Change-Id: I289d69f8a6bb8728c071d959dde74c7b2c0cfbf0