BUG#37023416: Make wait_condition.inc print the condition
* Background

Many mtr tests use the utility wait_condition.inc to wait on a
condition to occur.  When the condition does not occur within a
predefined timeout, it fails the test.  When replication is enabled,
it also prints useful debug info.

* Problem

But it does not print the wait condition itself. That is usually
useful to have, especially in tests that invokes wait_condition.inc
many times, and in particular if the test has changed recently so that
the line number in the output does not match the line number in your
tree.

(It may appear to print $wait_condition because it's included in the
message passed to `die`. However, `die` does not extrapolate the
message parameter, so it shows the dollar character and the variable
name rather than the variable value.)

* Fix

Make it print $wait_condition whenever it prints other debug info.

Change-Id: I9b1c874de3fdba36f7756408ed4222218b4cee3b