Bug#37023549 (1/5) Improve NdbApi SPJ scan protocol timeout handling
Add error insert to SPJ to allow a 'hard' timeout + failure to
close to be simulated.

Change-Id: Ie2479be2496ff4eb8ebe32547d8a8cb513708d26

Bug#37023549 (2/5) Improve NdbApi SPJ scan protocol timeout handling

Add SPJ scan testing

SPJ scans are added to the test as they have a different implementation
in the API and so need to be covered separately.

Due to bugs in SPJ, this test causes data node failure without
further fixes.

Change-Id: I1b7dd00a3bbaaa88aaf689de2ffc617afd041f62

Bug#37023549 (3/5) Improve NdbApi SPJ scan protocol timeout handling

Modify TC to disconnect bad APIs

TC detects in many cases when an API sends a signal which is
unexpected due the state discovered on an ApiConnectRecord.

Some cases are understood and expected, and others are not,
resulting in assertion failures in TC -> data node failure.

These can be useful for finding logic errors but can have
a large impact on users.

This patch modifies TC to handle some of these situations
differently :
 - Log information about the mismatch
 - Disconnect the API(s) involved
   (API sending the signal, API owning the ConnectRecord)

This will result in :
  - The API(s) being disconnected by all data nodes
  - API failure handling on all data nodes performing cleanup
    of the API's transaction objects, continuing with commit
    or rollback + then releasing them
  - APIs having to reinitialise their connection state to
    the cluster.

API disconnect + reconnect is generally less work, quicker
and less risky for system availability than a data node
restart.

This change in functionality is tested to some extent by
the SPJ timeout scenario in ndb_scan_protocol_timeout, but
other situations leading to state mismatch may exist.

Change-Id: Id88df14141983b4fef71e74ce275d6b4e26c6c58

Bug#37023549 (4/5) Improve NdbApi SPJ scan protocol timeout handling

On timeout, make SPJ API mark scans as needing close

SPJ API is aligned to the (fixed) behaviour of NdbApi on protocol
timeout handling - marking scans as needing close at the kernel.

This causes close() to attempt to close the scan, which will
succeed in cases where the scan has timed out due to LOAD.

Additionally, SPJ API is modified to log when these protocol
timeout handling paths are used, and what the result is.

The SPJ tests in ndb_scan_protocol_timeout for the 'LOAD' case
now pass with no leaks, but the BUG case needs a further fix.

Change-Id: I90e4421228f3db9c2de1de7ebb48ae81123abf68

Bug#37023549 (5/5) Improve NdbApi SPJ scan protocol timeout handling

SPJ Timeout Release on Close timeout

In the case where there is a timeout attempting to close an SPJ scan
at the kernel, this patch sets the ReleaseOnClose variable so that
the kernel side ApiConnectRecord will not be reused.

This aligns with the normal scan behaviour in the close-failed-timeout
case.

The ndb_scan_protocol_timeout testcase result is updated with the
expected results for SPJ scan timeouts due to load + bugs

Change-Id: I04b199759415c3e0d8f6f6546551c0271fac768c